+++ test/flushbar_test.dart
@@ -5,6 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+// ignore_for_file: deprecated_member_use_from_same_package
+
 void main() {
   test('Test Flushbar basic inicialization', () async {
     final flushbar = new Flushbar(message: "This is a test");